refactor: :bulb: optimize components directory
