+++ src/components/commons/AnimatedText/AnimatedText.tsx
