+++ src/components/layouts/PageLayout/PageLayout.tsx
@@ -1,7 +1,7 @@
 import type { FC, ReactNode } from 'react';
 
-import Footer from '../Footer/Footer';
-import Navbar from '../Navbar/Navbar';
+import Footer from '@/components/commons/Footer/Footer';
+import Navbar from '@/components/commons/Navbar/Navbar';
 
 type Props = {
   children: ReactNode;
